+++ webpack.config.js
@@ -1,10 +1,13 @@
 const path = require('path');
 
 module.exports = {
-    mode: 'production',
+    mode: 'development',
     entry: './src/gui/script.js',
     output: {
         filename: 'webview-bundle.js',
         path: path.resolve(__dirname, 'src', 'gui'),
+    },
+    optimization: {
+        minimize: false,
     },
 };